Updated NotificationSidebar file struckture
