+++ frontend/src/components/Notifications/NotificationCard/NotificationCard.scss
@@ -1,0 +1,52 @@
+@use "../../../resources/abstract/colors";
+@use "../../../resources/abstract/mixins";
+
+.notification-card {
+  width: 100%;
+  position: relative;
+  color: map-get(colors.$colors, 'textColor-primary');
+
+  padding: 1rem 0;
+  transition: all 200ms;
+
+  &:last-child {
+    border: none;
+  }
+
+
+  .content {
+    display: grid;
+    grid-gap: 8px;
+
+    .card-title {
+      font-weight: 400;
+      font-size: 16px;
+
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+
+      grid-gap: 10px;
+    }
+
+    .card-text {
+      color: map-get(colors.$colors, 'textColor-secondary');
+      font-weight: 300;
+      line-height: 1.4;
+      font-size: 13px;
+
+      a {
+        color: map-get(colors.$colors, 'textColor-primary');
+        font-weight: 400;
+      }
+    }
+
+    .timestamp {
+      color: map-get(colors.$colors, 'textColor-secondary');
+      font-weight: 400;
+      line-height: 1.4;
+      font-size: 12px;
+      white-space: nowrap;
+    }
+  }
+}
