+++ frontend/src/components/Sidebars/components/NotificationSidebar.scss
@@ -23,103 +23,9 @@
 
   .notifications {
     display: grid;
-    grid-gap: 0px;
+    grid-gap: 0;
 
     margin-top: 20px;
   }
 }
 
-.notification-card {
-  width: 100%;
-
-  position: relative;
-
-  color: map-get(colors.$colors, 'textColor-primary');
-
-  //border-left: 1px solid lighten(map-get(colors.$colors, 'background'), 5);
-
-  //background: darken(map-get(colors.$colors, 'background'), 3);
-
-  //border-radius: 5px;
-
-  padding: 1rem 0 1rem 1rem;
-  transition: all 200ms;
-
-  /*&:hover {
-    background: darken(map-get(colors.$colors, 'background'), 5);
-    cursor: pointer;
-
-    padding: 1rem;
-
-    transition: all 200ms;
-  }*/
-
-  &:before {
-    content: "";
-    width: 1px;
-    height: 100%;
-
-    position: absolute;
-    left: 0;
-    top: 0;
-
-    background: lighten(map-get(colors.$colors, 'background'), 10);
-  }
-
-  &:after {
-    content: "";
-
-    width: 8px;
-    height: 8px;
-
-    position: absolute;
-    left: -4.5px;
-    top: calc(1rem + 4px);
-
-    transform: rotate(45deg);
-
-    background: map-get(colors.$colors, 'background');
-    border: 1px solid lighten(map-get(colors.$colors, 'background'), 10);
-  }
-
-  &:last-child {
-    border: none;
-  }
-
-
-  .content {
-    display: grid;
-    grid-gap: 8px;
-
-    .card-title {
-      font-weight: 400;
-      font-size: 16px;
-
-      display: flex;
-      justify-content: space-between;
-      align-items: center;
-
-      grid-gap: 10px;
-    }
-
-    .card-text {
-      color: map-get(colors.$colors, 'textColor-secondary');
-      font-weight: 300;
-      line-height: 1.4;
-      font-size: 13px;
-
-      a {
-        color: map-get(colors.$colors, 'textColor-primary');
-        font-weight: 400;
-      }
-    }
-
-    .timestamp {
-      color: map-get(colors.$colors, 'textColor-secondary');
-      font-weight: 400;
-      line-height: 1.4;
-      font-size: 12px;
-      white-space: nowrap;
-    }
-  }
-}
